unit commit "form for brand"

+++ MyShopForHair.Core/Entities/Products.cs
@@ -7,7 +7,7 @@ namespace MyShopForHair.Core.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public ushort Price { get; set; }
         public int BrandId { get; set; }
         public Brand Brand { get; set; }
         public ICollection<Criteria> Criterias { get; set; }
